frontend/requirements.txt
fastapi>=0.115.0,<1.0.0
pikepdf>=9.0.0,<10.0.0
Pillow>=10.0.0,<11.0.0
vercel>=1.0.0,<2.0.0
